refactors: and use empty result
