+++ src/components/empty-message/index.tsx
@@ -1,20 +1,17 @@
 import { Image } from '@atb/components/icon';
 import { Typo } from '@atb/components/typography';
-import style from './empty-search-results.module.css';
+import style from './empty-message.module.css';
 
-export type EmptySearchResultsProps = {
+export type EmptyMessageProps = {
   title: string;
   details: string;
 };
 
-export default function EmptySearchResults({
-  title,
-  details,
-}: EmptySearchResultsProps) {
+export default function EmptyMessage({ title, details }: EmptyMessageProps) {
   return (
     <div className={style.container}>
       <div>
-        <Image image="OnBehalfOf" alt="" width="113" />
+        <Image image="EmptyIllustration" alt="" className={style.emptyImage} />
       </div>
       <div className={style.text}>
         <Typo.h3 textType="body__primary--bold">{title}</Typo.h3>
