+++ src/components/loading-empty-results/loading-empty-results.module.css
@@ -1,9 +1,3 @@
-.emptyImage {
-  display: block;
-  margin: 0 auto;
-  max-width: 15rem;
-}
-
 .isSearching {
   padding: 2.5rem;
   text-align: center;
