+++ src/translations/components/empty-search.ts
@@ -6,6 +6,24 @@ export const EmptySearch = {
     'You have not created a search yet.',
     'Du har ikkje gjennomført eit søk enno.',
   ),
+
+  emptyDetails: {
+    trip: _(
+      'Reiseforslagene dukker opp her når du søker etter det.',
+      'Travel suggestions will appear here when you search for it.',
+      'Reiseforslaga dukkar opp her når du søkjer etter det.',
+    ),
+    nearby: _(
+      'Avganger dukker opp her når du søker etter det.',
+      'Departures will appear here when you search for it.',
+      'Avgangar dukkar opp her når du søkjer etter det.',
+    ),
+    stopPlace: _(
+      'Avganger dukker opp her når du søker etter det.',
+      'Departures will appear here when you search for it.',
+      'Avgangar dukkar opp her når du søkjer etter det.',
+    ),
+  },
   searching: {
     trip: _(
       'Laster reiseforslag...',
